requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.2.3
SQLAlchemy<2.0
psycopg2-binary==2.9.6
rapidfuzz==2.13.7
openpyxl==3.1.2
gunicorn==20.1.0
requests==2.28.2
python-dotenv==0.21.0